factored out featured shows to separate file, added more shows;  caching posters; refactored services.py
